pyproject.toml
[project]
name = "django-base-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "ruff>=0.11.2",
]
